chore: separate SES from Pushwoosh email modules
