+++ src/domain/interfaces/email-api.interface.ts
@@ -1,4 +1,4 @@
-import type { CreateEmailMessageDto } from '@/modules/email/domain/entities/create-email-message.dto.entity';
+import type { CreateEmailMessageDto } from '@/modules/email/pushwoosh/domain/entities/create-email-message.dto.entity';
 
 export const IEmailApi = Symbol('IEmailApi');
