+++ src/modules/alerts/routes/alerts.controller.integration.spec.ts
@@ -11,8 +11,8 @@ import {
 import { IConfigurationService } from '@/config/configuration.service.interface';
 import type { Alert } from '@/modules/alerts/routes/entities/alert.dto.entity';
 import { EventType } from '@/modules/alerts/routes/entities/alert.dto.entity';
-import { EmailModule } from '@/modules/email/email.module';
-import { TestEmailApiModule } from '@/modules/email/datasources/__tests__/test.email-api.module';
+import { EmailModule } from '@/modules/email/pushwoosh/pushwoosh-email.module';
+import { TestEmailApiModule } from '@/modules/email/pushwoosh/__tests__/test.email-api.module';
 import { safeBuilder } from '@/modules/safe/domain/entities/__tests__/safe.builder';
 import {
   addOwnerWithThresholdEncoder,
