+++ src/modules/email/pushwoosh/datasources/pushwoosh-api.service.spec.ts
@@ -1,9 +1,9 @@
 import { FakeConfigurationService } from '@/config/__tests__/fake.configuration.service';
-import { PushwooshApi } from '@/modules/email/datasources/pushwoosh-api.service';
+import { PushwooshApi } from '@/modules/email/pushwoosh/datasources/pushwoosh-api.service';
 import { HttpErrorFactory } from '@/datasources/errors/http-error-factory';
 import { NetworkResponseError } from '@/datasources/network/entities/network.error.entity';
 import type { INetworkService } from '@/datasources/network/network.service.interface';
-import type { CreateEmailMessageDto } from '@/modules/email/domain/entities/create-email-message.dto.entity';
+import type { CreateEmailMessageDto } from '@/modules/email/pushwoosh/domain/entities/create-email-message.dto.entity';
 import { DataSourceError } from '@/domain/errors/data-source.error';
 import { faker } from '@faker-js/faker';
