+++ src/modules/email/pushwoosh/datasources/pushwoosh-api.service.ts
@@ -5,7 +5,7 @@ import {
   INetworkService,
   NetworkService,
 } from '@/datasources/network/network.service.interface';
-import { CreateEmailMessageDto } from '@/modules/email/domain/entities/create-email-message.dto.entity';
+import { CreateEmailMessageDto } from '@/modules/email/pushwoosh/domain/entities/create-email-message.dto.entity';
 import { IEmailApi } from '@/domain/interfaces/email-api.interface';
 
 @Injectable()
