+++ src/modules/email/pushwoosh/pushwoosh-email.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { HttpErrorFactory } from '@/datasources/errors/http-error-factory';
 import { IEmailApi } from '@/domain/interfaces/email-api.interface';
-import { PushwooshApi } from '@/modules/email/datasources/pushwoosh-api.service';
+import { PushwooshApi } from '@/modules/email/pushwoosh/datasources/pushwoosh-api.service';
 
 @Module({
   providers: [HttpErrorFactory, { provide: IEmailApi, useClass: PushwooshApi }],
